tools/evaluation/robotcar/sdk_python/requirements.txt
numpy
matplotlib
colour_demosaicing
pillow
opencv-python
open3d-python
